+++ bee-message/CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Example that counts the number of allocations while scoring PoWs
-
 ### Changed
 
 ### Deprecated
@@ -22,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security -->
 
 ## 0.2.0 - 2022-XX-XX
+
+### Added
+
+- `ByteCost` and storage deposit verification;
 
 ### Changed